feat: isPrincipal instead of instanceof for json principal (#1038)
# Motivation

While migratign Juno to AgentJS v3
([PR](https://github.com/junobuild/juno/pull/1940)), one of my test
started failing as the assertion `toBeInstanceOf(Principal)` returned
false. I assume this happened because the `Principal` instance used in
the test was not the same as the object returned through my library. To
overcome the issue, I migrated the test to `Principal.isPrincipal`.
That's why, I think we should enhance the `jsonReplacer` with same
assertion to avoid those kind of issue.

# Notes

~~I assume none will be affected by the changes, thinking that aside
from NNS dapp and Juno there are rare consumers of the `jsonReplacer`,
but still, labelled as breaking change because clients which in the past
would have serialized principal not instead of will start serializing
those as principal, which is per sé not incorrect.~~

After thinking at it, let's release it as a feature. It's a breaking
only if the outcome is saved in a storage and I'm probably the only one
who do that. Plus, it's not a breaking per sé as we can consider that
the detection was not precise enough and it can be an improvement or
fix. Also easier to ship in our dapps in which it is not a breaking.

# Changes

- Use `Principal.isPrincipal` instead of `instanceof Principal` in
`jsonReplacer`

---------

Signed-off-by: David Dal Busco <david.dalbusco@dfinity.org>
Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>