refactor: extract snapshot interfaces and mapper utils (#1052)
# Motivation

This PR makes the main API easier to read - less code and types
duplication - and also initialize a new dedicated modules for the params
related to snapshot. Useful as we will have to add more types to support
download and upload snapshots.

# Changes

- Init new module `snapshot.params.ts`
- Move snapshot related interface to it
- Create an `OptionSnapshotParams` and use it to type
`takeCanisterSnapshot` (instead of using TS utils in the main API)
- Create and use two utilities to map the snapshot params (to avoid code
duplication)

---------

Signed-off-by: David Dal Busco <david.dalbusco@dfinity.org>
Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>