chore: upgrade didc v0.5.1 (#1068)
# Motivation

Before migrating to `@icp-sdk/bindgen`, we need to upgrade `didc` and
regenerate the types; otherwise, we risk mixing declaration differences
during the migration.

This is especially relevant for the TypeScript definitions of functions,
used in e.g. `ledger-icrc`, which were modified in this [Candid
PR](https://github.com/dfinity/candid/pull/627).

It's also interesting to upgrade because latest version of `didc` parses
the comments to the output files as well.

# Changes

- Upgrade `didc` to latest version `v0.5.1` published in Candid release
[2025-08-04](https://github.com/dfinity/candid/releases/tag/2025-08-04)

Signed-off-by: David Dal Busco <david.dalbusco@dfinity.org>