feat: generate ledger-icrc candid declarations with didc v0.5.1 (#1077)
# Motivation

Generate the DID declarations with didc `v0.5.1`.

# Notes

The auto-generated comment added at the top of the DID file might be
misinterpreted by didc as a comment for a type.
This has been patched in Candid PR
[681](https://github.com/dfinity/candid/pull/681) which we will
integrate once [v0.5.2](https://github.com/dfinity/candid/pull/682) is
released.

You might noticed that a `callback` type isn't typed anymore as
`QueryArchiveFn` (a function) but become a `[Principal, string]`.
According SDK team this is expected and is the result of this Candid fix
[PR](https://github.com/dfinity/candid/pull/627).

# Changes

- Copy declarations from the PR #1069 that was generated with the CI job

Signed-off-by: David Dal Busco <david.dalbusco@dfinity.org>