feat: snapshot read metadata response type and mapper (#1079)
# Motivation

When I developed support for the read and upload snapshot feature in
#1046, I was really hesitant about whether to parse/map the metadata
response. On one hand, for consistency with the other functions in the
module, we should not parse the response, since none of the other
functions do so. On the other hand, we already use custom parameters for
the exposed parts of the module, a pattern that should also apply when
implementing the upload metadata of the snapshot.

Ultimately, I decided it would be more practical for the consumer to
receive a mapped object when reading data. This way, the consumer can
simply download a value and pass the same back on upload, without having
to implement additional conversions on their side.

I validated this approach with my implementation in Juno's CLI (PR
[382](https://github.com/junobuild/cli/pull/382)).

That said, I’m still not fully satisfied with it. Generally speaking, I
hope we’ll find a solution in the future that makes both inputs and
outputs consistent.

# Changes

- Define a response type to read metadata
- Add a utility to parse candid types to custom js type
- Expose new module to consumer

# Notes

I exposed the all `snapshot.response` module in `index.ts`, not just the
interface, for consistency reasons because we do so with the params.

---------

Signed-off-by: David Dal Busco <david.dalbusco@dfinity.org>