feat: assertNever utils for unreachable variant type (#1081)
# Motivation

A cool idea of @AntonioVentilii shared in this
[review](https://github.com/dfinity/ic-js/pull/1079#discussion_r2400960151):
in addition to throwing when a variant is unknown, which can happens
when the did types are extended in the IC repo, we can also type the
rest to detect the issue at build time. Given there are various usage of
the pattern in ic-js, I thought that creating a reusable function can be
handy.

# Notes

I don't know how to actually implement a vitest that would passes the
build with an incompatible code. Thought it was kind of a quite an
effort so skipped that part.

That said we can assert it works out as implementing the function in nns
allows to detect that missing types are not implemented, see #1083.

# Changes

- Create `assertNever`

---------

Signed-off-by: David Dal Busco <david.dalbusco@dfinity.org>
Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>