feat: replace usage of Buffer in nns-js to convert uint8Array to hex (#1088)
# Motivation

We want to remove the use of `Buffer`. While testing the integration in
NNS-dapp ([PR](https://github.com/dfinity/nns-dapp/pull/7348)), @yhabib
noticed that the application was complaining about the missing
polyfilled `Buffer`. This is likely because `nns-js` still relies on it
for two features.

# Changes

- Replace usage of `Buffer` with `uint8ArrayToHexString` from
`@dfinity/utils`
- Updated also in a comment in the unlikely event we would uncomment it
in the future, just to avoid surprises

Signed-off-by: David Dal Busco <david.dalbusco@dfinity.org>