feat: add to account disburse maturity (#1094)
# Motivation

Support the parameter `toAccount` in the function to disburse maturity
in the NNS package.

I want to add and test the functionality of all the parameters in the
Ledger hardware wallet.

# Changes

* Updated `toDisburseMaturityRequest` in `request.converters.ts` to
accept an optional `toAccount` parameter and correctly serialize it for
the disbursement command.
* Modified the `disburseMaturity` method in `governance.canister.ts` to
accept the new `toAccount` parameter.
* Updated documentation for `disburseMaturity` to describe the new
`toAccount` parameter and its behavior.

# Tests

* Added a test case in `governance.canister.spec.ts` to verify that
maturity can be disbursed to an ICRC account, including correct
serialization and service invocation.

# Todos

- [X] Add entry to changelog (if necessary).

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>