build(core): modernize build (node esm and pkg exports) (#1108)
# Motivation

The build pipeline is a bit outdated. For NodeJS we are shipping CJS and
our `package.json` do not use the `exports` field to declare modules.
Now it's a good time to modernize this.

# Notes

- The setup is similar as the one I've been using in
[juno-js](https://github.com/junobuild/juno-js) since months, if not
years.
- `nns-proto` requires CJS for Node. If ESM is used, the build shows
warning like `Import "BallotInfo" will always be undefined because the
file "proto/governance_pb.js" has no exports [import-is-undefined]`.
Given no one I think is using this library anymore, we just keep CJS for
backwards compatibility instead of trying to modernize it.

# Changes

- Migrate NodeJS build from CJS to ESM (except for proto)
- Add `exports` field to `package.json`
- Update `dist` structure from `esm` and `cjs` to `browser` and `node`
subfolders.
- Various minimal clean-up in `esbuild` script (ignore mock files, avoid
peer dependencies code duplication)
- Specify `"type": "module",` to each `package.json`

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>