feat(canisters): init @icp-sdk/canisters (#1106)
# Motivation

Similarly to `@icp-sdk/core` or `@icp-sdk/auth`, we want to ship a
modular library that provides the features for interfacing with
canisters on the Internet Computer.

# Notes

This PR provides the standard setup but, it will be modified in next PRs
because the library isn't a single entry package but will be a module
entry-point pattern.

# Changes

- Setup workspace package, package.json, README, license etc.