build(core): migrate agent-js to icp-sdk (#1104)
# Motivation

We want to migrate agent-js to `@icp-sdk/core`. Since this library
currently re-export all underlying libraries - e.g. `@dfinity/agent` and
`@dfinity/principal` - we are using, in this first step we can solely
switch the dependencies.

# Temporary "Workaround"

Without migrating all imports to their corresponding sub-paths (will be
resolved in #1044), the generate bundles end including the code of the
agent which per extension lead the library to become more heavy. As
temporary workaroung, we add an explicit exclude in the bundler script
which we can remove afterwards.

# Notes

Announcement:
https://forum.dfinity.org/t/icp-javascript-sdk-v4-0-0-is-released/55915

# Changes

- Update helper script that handles lib upgrades.
- Replace `@dfinity/*` peer dependencies by `@icp-sdk/core`.
- Exclude `@icp-sdk/core` from `size-limit` checker (similarly to the
`@dfinity/...` libs).
- Temporary manually list `@dfinity/...` as peer dependencies for the
build (see workaround above).

---------

Signed-off-by: David Dal Busco <david.dalbusco@dfinity.org>