feat(core): migrate agent-js to icp-sdk (#1044)
# Motivation

While we can technically continue to import modules from AgentJS using
the legacy libraries - inherited as peer dependencies of `@icp-sdk/core`
- we want to fully migrate to the new subentry-paths library. That is
why this PR adapts all imports. Basically a search and replace.

# Notes

This PR does not migrate the imports in the DID files and the scripts
that generates those. This will be resolve separately and partially -
for the huge part - automatically.

# Changes

- Search `@dfinity/agent` and replace by `@icp-sdk/core/agent`
- Search `@dfinity/identity` and replace by `@icp-sdk/core/identity`
- Search `@dfinity/candid` and replace by `@icp-sdk/core/candid`
- Search `@dfinity/principal` and replace by `@icp-sdk/core/principal`

---------

Signed-off-by: David Dal Busco <david.dalbusco@dfinity.org>
Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>