feat(canisters): icp-sdk/canisters should reexport ckbtc (#1111)
# Motivation

The `@icp-sdk/canisters` is meant to reexport the existing DFINITY
libraries related to canisters. This PR provides the re-export for
`@dfinity/ckbtc`.

# Notes

The peer dependency is added with a pinned requirement. Given that we
are going to release both DFINITY and `icp-sdk/canisters` libraries in
parallel, we want to enforce devs using the latest to inherit what's
actually shipped and specified in the CHANGELOG.

# Changes

- Add peer dependencies.
- `export * from ...`
- Remove files that become unused following this PR

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>