chore(repo): build icp-sdk/canisters before test (#1116)
# Motivation

The `@icp-sdk/canisters` need to be build before running tests. In the
CI we already do but locally not. So this PR adds the npm script entry
to build before test.

# Notes

In the CI it means this lib will be build twice when running the tests
but, I guess it's acceptable for now?

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>