feat(canisters): icp-sdk/canisters should reexport ledger-icp (#1129)
# Motivation

The `@icp-sdk/canisters` package is meant to re-export the existing
DFINITY libraries related to canisters. This PR provides the re-export
for `@dfinity/ledger-icp`.

# Design 🤔

For this particular library, I was hesitant about exposing it from
`@icp-sdk/canisters/ledger-icp` or `@icp-sdk/canisters/ledger/icp`,
given that we also have an ICRC ledger.

Ultimately, I went with the former because it felt more consistent and
followed the historical pattern. That said, I’m not fully convinced, if
anyone thinks the latter is better, I’m happy to change it.

# Notes

The peer dependency is added with a pinned requirement. Given that we
are going to release both the DFINITY and `@icp-sdk/canisters` libraries
in parallel, we want to enforce that developers use the latest version
to inherit what is actually shipped and specified in the CHANGELOG.

# Changes

- Add peer dependencies  
- `export * from ...`