feat(canisters): icp-sdk/canisters should reexport sns (#1132)
# Motivation

The `@icp-sdk/canisters` package is meant to re-export the existing
DFINITY libraries related to canisters. This PR provides the re-export
for `@dfinity/sns`.

# Notes

The peer dependency is added with a pinned requirement. Given that we
are going to release both the DFINITY and `@icp-sdk/canisters` libraries
in parallel, we want to enforce that developers use the latest version
to inherit what is actually shipped and specified in the CHANGELOG.

# Changes

- Add peer dependencies  
- `export * from ...`