build(core): downgrade @icp-sdk/bindgen to v0.1.0-beta.6 (#1125)
# Motivation

When I prepared the migration from `didc` to `@icp-sdk/bindgen`, I
tested the latter library and ensured that, aside from the new
declarations parsed at the root of the DID modules, there would be no
other changes. The goal of the next release was to update the tooling
without introducing any code changes. I prepared this migration using a
beta version.

However, unfortunately, it turns out that a breaking change was added to
the version of `@icp-sdk/bindgen` released as v1, which was not included
in the betas I had tested.

As shown in #1124, various types are now declared without unions.

`didc`:

```
export interface Account {
  owner: Principal;
  subaccount: [] | [Uint8Array | number[]];
}
```

`bindgen`:

```
export interface Account {
  owner: Principal;
  subaccount: [] | [Uint8Array];
}
```

# Changes

Given that I maintain that upgrading the tooling should be fully
backward compatible and should not impact the code in any way, I propose
to downgrade `@icp-sdk/bindgen` for the time being. Once we have
migrated and released, we can then upgrade to a newer version