chore(core): remove test-vectors scripts (#1135)
# Motivation

Following the modernization of the build (NodeJS cjs -> esm) and
migration to `@icp-sdk/core`, the question arised: what to do with the
`test-vectors` scripts?

Given that those are not maintained by the OISY and NNS Dapp team -
owners of this repo - and in accordance with @lmuntaner, we decided to
move those elsewhere respectively and probably in the
https://github.com/dfinity/hardware-wallet-cli/ repo. Llorenç will take
care of this (when he has a bit of time) but, we can already remove
those from this repo given that we are aiming to release a major version
soon.

# Changes

- `rm - r ./scripts/test-vectors`