feat(canisters): reexport ledger icrc and icp under same parent (#1141)
# Motivation

After discussing with @AntonioVentilii we concluded that exposing both
ledgers ICP and ICRC under a same parent folder in the multy entry paths
library was a better solution for readability and communication. It's
also to some extension more future prone as it follows a pattern
`@dfinity/canisters/{canister}/something`. For example this allows us
someday (or not) to move out the Index canister from the ledger.

In short, this PR does:

`import * from @icp-sdk/canisters/ledger-icp` -> `import * from
@icp-sdk/canisters/ledger/icp`
`import * from @icp-sdk/canisters/ledger-icrc` -> `import * from
@icp-sdk/canisters/ledger/icrc`

# Notes

- Follow-up of #1130 and #1129
- I tried to move the related TypeScript modules and test to a similar
files structure (e.g. `packages/canisters/src/ledger/icp.ts`) but, this
is not supported by the build which looks for entry point at the root of
the `src` folder (`packages/canisters/src`). I'm afraid that starting
searching for entry point recursively will lead to issue in the future
therefore decided to not move the modules given that anyway it's the
`package.json` definition that matters for the consumer.

# Changes

- Change reexport in `package.json` from `ledger-icp` and `ledger-icrc`
to `ledger/icp` respectively `ledger/icrc`
- Update test with new lazy loading import path