build(canisters): split subpaths structure (#1144)
# Motivation

We want to embed the generated JavaScript code in `@dfinity/canisters`,
along with the corresponding TypeScript declaration files. As a result,
we also need to include the Candid `.d.ts` files in the distributed
library.

Currently, all JavaScript and TypeScript files are flattened at the root
of the library. This would require flattening the Candid files as well,
which could lead to naming clashes.

To avoid this, we need to refactor the library to use a folder-based
structure. This way, each subentry can contain its own `candid` folder,
keeping related files organized and preventing conflicts.


# Changes

- Move reexported modules to a subfolders based structure
- Rename TS file and test to `index`
- Update `esbuild.mjs` to discover what entry points to build based on
the `exports` field of the `package.json` of the subpaths-only libraries

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>