build(canisters): copy candid from dfinity libs over (#1145)
# Motivation

We want to include the Candid declarations and files in the
`@icp-sdk/canisters` library. Since we plan to release both the legacy
libraries and the new library in parallel - with the legacy ones
continuing to be developed - we copy the Candid files after the build
process.

# Changes

- Add `copy-candid.mjs` script which also look for files to copy based
on the `exports` field of `package.json`
- Set up git exclusion for the new scripts

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>