chore(core): generate declarations from reexport libs for multi-path lib (#1164)
# Motivation

And one more issue... So, if we generate the TypeScript declarations
(`.d.ts`) with `tsc` as we generally do, given that the re-exported
libraries in the multi-path library are solely exported with a
one-liner, we end up with a declaration that is also a one-liner.

That’s a problem, as we aim to embed the code — not re-export the legacy
libraries (which we’ll continue to develop in parallel, yolo).

In other words, without any particular action, if we have the following
in the TypeScript file:

```
export * from '@dfinity/ckbtc';
```

We get the following in the related `.d.ts`:

```
export * from '@dfinity/ckbtc';
```

Unfortunately, there’s no `tsc` option that would “inline” the types.
Therefore, we have no other choice than to add another hack, which
consists of generating the types for each source library but targeting
the multi-entry library.

# Notes

In the end, I wonder if a simple copy/paste of the ckbtc/dist folder
from the source to canisters/ckbtc wouldn’t be easier but well, here we
are.

# Changes

- Create a new post build script that does the above.