chore(core): size-limits for icp-sdk/canisters (#1167)
# Motivation

We want indicators about the size of `@icp-sdk/canisters`.

# Notes

The size differences in NNS and SNS between the legacy and new libraries
are, I believe, due to the fact that excluding ledger-icrc and
ledger-icp from the size comparison works for the legacy version but not
for the new one, because those dependencies are no longer peers and end
up being built into separate chunks - i.e. the size checker does not
notice it should exclude the chunk size.

# Changes

- Copy current setup for legacy libs
- Adapt naming and paths