build(core): use @icp-sdk/bindgen v0.2.0-patch-typed-arrays (#1171)
# Motivation

We want to migrate from `didc` to `@icp-sdk/bindgen` without introducing
any changes in the generated DID files, except for updating the imports
from `@dfinity/...` to `@icp-sdk/core`.

In previous PRs, we noticed two issues:

- `bindgen` parses the types as `Uint8Array` instead of `Uint8Array |
number[]`.
- It also duplicates the DID types at the root of the modules (which is
great for backend development) but makes the code heavier.

The latter has been patched in
[v0.2.0](https://github.com/dfinity/icp-js-bindgen/blob/main/CHANGELOG.md#020-2025-10-20)
and has become an opt-in feature, which we don’t necessarily need to use
right now.

The former is now the default behavior, which we can start using once
we’ve migrated (it’s a breaking change).

That’s why we’re using a patched version of `@icp-sdk/bindgen`, released
to address both concerns.