chore(core): set version for next release (#1174)
# Motivation

While fundamentally there are no features or fixes in the next release,
we want to make developers aware there are important changes underneath
as the libraries are migrated from `@dfinity/...` to `@icp-sdk/core`
peer dependencies.

That is why we bump the libs to a major version.

# Notes

We want to release first a `@next` version and test this in OISY and NNS
dapp. We also want to test the new `@icp-sdk/canisters`. That's why this
PR does not bump the version number of the root package nor provide the
CHANGELOG yet.

# Changes

- Bump major version in legacy dependencies and in utilities.