build(core): set access public to publish next to support new libs (#1177)
# Motivation

The publish `@next` of `@icp-sdk/canisters` failed (see
[job](https://github.com/dfinity/ic-js/actions/runs/18656091303/job/53185451080))
because unlike our main publish script, the "next" did not contains the
keywords `--access public`. I guess it's because we assumed that next
was always published after a first public release but here we are trying
to released a flagged library first.

# Changes

- Add `--access public` to publish next action scripts