chore(core): pre- and post-publish scripts for icp-sdk/canisters (#1183)
# Motivation

> By default, npm install will install all modules listed as
dependencies in package.json(5).

Source: https://docs.npmjs.com/cli/v8/commands/npm-install

Which means that by default, even though we bundle the code within
`@icp-sdk/canisters`, the devs will still have to install the
dependencies which is not that an issue. However, as I'm trying to use
the `@next` version in NNS dapp, I cannot proceed, I'm getting errors:

> npm error code ETARGET
> npm error notarget No matching version found for @dfinity/ckbtc@^5.

This happens because the depencies major versions are not yet published.
In other words, even though we do not need the libs, they can become an
issue for installation.

That is why we need to remove the dependencies from the `package.json`
we ship to npm.

# Notes

I followed the same approach as we already have in other build scripts.
Took a shortcut for the pre-script by just removing according a key and
not the `exports` field. Felt more straightforward.

Anyway we want to move away from those workaround sooner than later.

# Changes

- Create a pre-publish scripts that remove the `dependencies` from the
`package.json`
- Similarly, create a post-publish script that redo the `dependencies`.