feat(sns): expose missing did types (#1185)
# Motivation

Not sure but I believe it's due to the new `bundler` target for
TypeScript, when consumed, the consumers cannot directly used the types
available in the candid subfolder of the libraries - regardless if
legacy or new multi paths.

I don't really know how to resolve it and if it's worth spending hours
on the topic given that the libraries already expose the DID types that
are suppose to be visible by the consumers.

Furthermore or anyway, I think we should migrate to the pattern I
started using in Juno: exporting with a root "namespace" the did types.

# Notes

To assert this PR works out I released a next version from the branch
and upgrade the libs in https://github.com/dfinity/nns-dapp/pull/7453.
Build and linter are now working there.

# Changes

- Anyway, this PR exposes the missing types required to update the
`@next` libraries in NNS dapp (see PR
https://github.com/dfinity/nns-dapp/pull/7453)