chore(canisters): fix publishing scripts (#1191)
# Motivation

Despite the rules, the build and publish scripts of `@icp-sdk/canisters`
are still published to npm. However, now that we are removing the root
`.npmignore` (see #1190), we can change the approach. Instead of telling
npm what we want to publish and trying to provide a pattern to exclude
some, we can tell npm we want to publish everything except few things.

# Changes

- Remove `files` from package.json
- List ignore files in `.npmignore` of the lib