chore(core): remove useless .npmignore (#1190)
# Motivation

As far I can tell, the root `.npmignore` file has absolutely no effect
in this repo. e.g. we can publish `packages/canisters` even though
according this files few data would be ignored.

I think it's because we use `files` in our `package.json` which takes
precendence of this ignore file.

Therefore, let's remove it.

**Update**

We also need to remove it for #1191.