docs(canisters): the library does not reexport legacy libs as peer anymore (#1205)
# Motivation

While reviewing before release, I just noticed that a mention in
`@icp-sdk/canisters` was outdated since we switched approach along the
way - i.e. it does not ship the legacy libs as peer but bundle them.

I also added a plural to `Example` - since I was there - because there
are two examples.