fix(ic-management): enforce update call for canister status (#1218)
# Motivation

The new DID files introduced in
[#1209](https://github.com/junobuild/juno/pull/1209) added
`canister_status` as a **query** call.

At first glance, this might not seem problematic. However, the
`ICManagementCanister` was originally implemented to support only
**update** calls and relied on the `idlFactory` to initialize the
service.

Until now, this worked fine because the uncertified `idlFactory` was
effectively the same as the certified one (`service` vs.
`certifiedService`). But with this recent change, when `canisterStatus`
is invoked, it uses `service`, which now means it attempts a **query**
call, not an **update** anymore.

The problem arises because the local replica (in PocketIC v10) still
supports **only update** calls - it’s not necessarily up to date with
this new behavior.

In short, as you can see in the failed Juno build ([GitHub Actions
log](https://github.com/junobuild/juno/actions/runs/19163678544/job/54779499282?pr=2211)),
this change has introduced a bug and caused pipelines to break.

# Notes

I'm not entirely sure this is the root cause of the issue. I wanted to
release a version from a branch to assert the solution but, it failed
https://github.com/dfinity/icp-js-canisters/actions/runs/19167647290. As
I'm blocked and absolutely need a fix for yesterday, plus the fact that
it PR is fundamentally correct, I propose that we move on with review
and merge on main once approved.

# Changes

- Use `certifiedService` instead of `service` in canister implementation
- Keep `service` for implementation of query in next PR

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>