feat(ic-management): fetch_canister_logs must be a query call (#1220)
# Motivation

So guess what, the Juno pipeline failed on another issue 😄. Turns out
`fetch_canister_logs` must be a query when called e.g. from a NodeJS
script or a frontend app.

See the test failing here:
https://github.com/junobuild/juno/actions/runs/19168787012/job/54796088776?pr=2211

I confirmed it with the interface spec team. Until today it worked out
because there was only a single type of service implemented in
`IcManagementCanister` and unlike what was implemented in #1218, this
particular call must use the idl factory for query.

# Changes

- Use the `this.service` instead of `this.certifiedService` for
`fetchCanisterLogs`

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>