chore(release): reference dependencies with next instead of wildcard (#1227)
# Motivation

Referencing any version when publishing `next` seems to be an issue when
releasing the new multi-entry library `@icp-sdk/canisters`. It seems
that instead of requiring the latest `next` version, it requires the
most recent one, which can be the one effectively released.

That is why this PR replaces the references from `@dfinity` when
releasing `next` (or `beta`) with the tag.

# References

https://docs.npmjs.com/cli/v11/configuring-npm/package-json#dependencies

> See [semver](https://github.com/npm/node-semver#versions) for more
details about specifying version ranges.
> ...
> `tag` A specific version tagged and published as `tag` See [npm
dist-tag](https://docs.npmjs.com/cli/v11/commands/npm-dist-tag)

# Changes

- Update `update-version.mjs` to set appropriate references.