refactor(build): Move repo details in `import-candid` script (#1238)
# Motivation

The function `import_did` inside `import-candid` script uses global
values for details of the source repo, derived from the global variable
`IC_DIR`.

However it even accepts as input the source directory that just falls
back to `IC_DIR` in case not provided.

So, to have a more consistent logging, we move the definitions of the
repo details right after the definition of the source directory.