chore(release): prepare v85 (#1261)
# Motivation

We want to release a version with the latest changes.

# Notes

Most libs has been pumbed to major due to the IDL change, however, I did
not do so in NNS because Candid is obbfuscated there with mapper and
interfaces and while I did bump it's reference to ledger-icp, felt like
it was not a major in this case.

# Changes

- Set versions and prepare CHANGELOG