feat(ledger): Add ICRC-10 did files (#1282)
# Motivation

The standard ICRC-7 must implement method `icrc10_supported_standards`
(according to
[specifications](https://github.com/dfinity/ICRC/blob/main/ICRCs/ICRC-7/ICRC-7.md#icrc10_supported_standards)).
However, that is not in the `.did` file of the ICRC-7 standard (see
[here](https://github.com/dfinity/ICRC/blob/main/ICRCs/ICRC-7/ICRC-7.did)).

On the contrary, for ICRC-1, we have it in the same candid file because
we use the `.did` file of the ledger suite (see
[here](https://github.com/dfinity/ic/blob/master/rs/ledger_suite/icrc1/ledger/ledger.did)).

Now, I need to implement a method that calls
`icrc10_supported_standards` in the new `IcrcNftLedgerCanister` class.

The plan is to create another abstract class, that is purely
`Icrc10Canister` (or something similar) that uses the `.did` file of
ICRC-10 (see
[here](https://github.com/dfinity/ICRC/blob/main/ICRCs/ICRC-10/ICRC-10.did)).

This class will extend the existing abstract class `Canister`.

Then, I make `IcrcLedgerCanister` and `IcrcNftLedgerCanister` implement
`Icrc10Canister` instead of just `Canister`.

As first step in this PR, we add the candid files (and bindings) of
standard ICRC-10.

# Changes

- Include ICRC-10 did file in the script to import candid files.
- `npm run did`

# Tests

No tests required yet.