chore(fix): beta version reference (#1289)
# Motivation

I did not tried it out but, I read the code of `update-version.mjs`
which we modified this week to ship `@icp-sdk/canisters@next` and I
think we - or I - made a small "mistakes". We support `next` and `beta`
version as well, therefore the tag should be use as reference not `next`
hardcoded.