fix(core): download ICRC-1 unsupported URL (#1290)
# Motivation

The URL provided in #1287 to download the ICRC-1 did files is actually
incompatible with the `import-candid` script.
It leads to an error:

```
jq: error (at <stdin>:5): Cannot index object with number
Failed to retrieve commit hash for main/standards/ICRC-1/ICRC-1.did in dfinity/ICRC-1.
```

As seen in this week automatic job that ended in error:
https://github.com/dfinity/icp-js-canisters/actions/runs/19623933541/job/56189246360

# Changes

- Adapt URL to download did file